// src/core/api/generated/model/getApiForumThreadsByThreadId200ThreadAuthor.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Vewave API
 * Backend API for Vewave watch-together rooms and playback synchronization.
 * OpenAPI spec version: 1.0.0
 */

/**
 * @nullable
 */
export type GetApiForumThreadsByThreadId200ThreadAuthor = {
  id: string
  name: string
  /** @nullable */
  username: string | null
  /** @nullable */
  avatarUrl: string | null
} | null